+++ extensions/pocket/src/latest.ts
@@ -1,24 +1,23 @@
 import { open, Toast } from "@raycast/api";
 import { fetchBookmarks } from "./utils/api";
 import { HTTPError } from "got";
-import { ReadState } from "./utils/types";
 
-export default async function() {
+export default async function () {
   const toast = new Toast({
     title: "Searching bookmarks",
-    style: Toast.Style.Animated
+    style: Toast.Style.Animated,
   });
 
   toast.show();
 
   try {
-    const bookmarks = await fetchBookmarks({ state: ReadState.Unread });
+    const bookmarks = await fetchBookmarks({ count: 1 });
 
     if (bookmarks.length === 0) {
       toast.style = Toast.Style.Failure;
       toast.title = "No bookmarks found";
     } else {
-      const bookmark = bookmarks[Math.floor(Math.random() * bookmarks.length)];
+      const bookmark = bookmarks[0];
       await open(bookmark.pocketUrl);
     }
   } catch (error) {